Switch index to TS
