--- src/pages/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const IndexPage = () => {
-  return (
-    <main>
-      <title>Jonathan Horn, Web Developer</title>
-      <h1>Jonathan Horn</h1>
-      <h2>Full Stack Web Developer</h2>
-    </main>
-  );
-};
-
-export default IndexPage;
