+++ src/pages/index.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Layout from '../components/layout';
+
+const IndexPage = () => {
+  return (
+    <Layout>
+      <main>
+        <title>Jonathan Horn, Web Developer</title>
+        <h1>Jonathan Horn</h1>
+        <h2>Full Stack Web Developer</h2>
+      </main>
+    </Layout>
+  );
+};
+
+export default IndexPage;
